+++ test/testIssueLocation.py
@@ -1,21 +1,6 @@
 from unittest import TestCase
 from bot import IssueLocation
-
-
-class DataStream:
-	def __init__(self):
-		self.data = []
-
-	def write(self, **args):
-		assert len(args) == 1
-		k, v = args.popitem()
-		if hasattr(v, 'print'):
-			ss = DataStream()
-			v.print(ss)
-			value = ss.data
-		else:
-			value = v
-		self.data.append((k, value))
+from test.DataStream import DataStream
 
 
 class TestIssueLocation(TestCase):
